Modified blueprints.xml.append
I've resisted the urge to go ham on this as best as possible to offer as many suggestions as is feasible. Most often in the Short Names, but there is also a suggestion for altering the names of each Trade Good tier from (Very Basic/Basic/Common/Valuable/Rare) to (Basic/Common/Uncommon/Rare/Very Rare) to make value distinctions between the three bottom tiers much easier to ascertain in the eyes of those new to CE.

On a side-note, this site is so confusing to try and use, nothing really makes sense. But the text editor is absolutely garbage, it was taking so long for it to respond to even the slightest text changes that simply copying it into Notepad++ and editing it there shaved the editing time in half.